Weather_App/requirements.txt
Flask==3.0.3
requests==2.32.0
geopandas==1.0.1
Shapely==2.0.6
python-dotenv==1.0.1
gunicorn==22.0.0 
prometheus_client ==0.21.0